OpenSaleTemplate.xhtml Added navbar across the top. This serves two purposes: Cosmetic improvement, tells you who is logged in, and has a logout button. Added resolution scaling objects on all pages. You can test this by resizing any page in the software and noting how it moves the content to your view. http://quirktools.com/screenfly/ - Use this to show how pages appear with accepted resolutions.
catalogModify.xhtml
Renamed catalog.xhtml to this as we have both viewCatalog.xthml and catalogModify.xhtml pages.  Corrected information on this page and made link to editProduct.xthml.

editProduct.xhtml
Created. Untested and needs alignment.

editUser.xhtml
Corrected naming convention.

index.html
modified inputText fields to appropriately size on scaled

mainMenu.xhtml
Redesigned with simplicity. Not final. Only shows fields that are appropriate for your level and will display them logically. Happy with the functionality but not the design. 

stylesheet.css
Made changes as needed across page elements. 

userList.xhtml
Removed redundant features and features that were not working.  Resized and formatted it properly.

viewCatalog.xhtml
This exists because we give management the ability to only view the catalog. We could remove this and just hide the functionality for the permissions. Up to you guys.